Format rune supply using divisibility (#2509)

+++ src/runes/etching.rs
@@ -2,6 +2,6 @@ use super::*;
 
 #[derive(Default, Serialize, Debug, PartialEq, Copy, Clone)]
 pub struct Etching {
-  pub(crate) divisibility: u128,
+  pub(crate) divisibility: u8,
   pub(crate) rune: Rune,
 }
